requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
pillow
pytesseract
moviepy
speechrecognition
pydub